Implemented SetContextCulture attribute. A bit of refactoring
