+++ src/Atomic.Common/Api/Preview/EnablePreviewRouteAttribute.cs
@@ -1,0 +1,6 @@
+﻿namespace Atomic.Common.Api.Preview
+{
+    [AttributeUsage(AttributeTargets.Class, AllowMultiple = false, Inherited = true)]
+    public sealed class EnablePreviewRouteAttribute : Attribute
+    { }
+}
