--- src/Atomic.Common/Api/PreviewRoute/PreviewRouteAttribute.cs
@@ -1,6 +1,0 @@
-﻿namespace Atomic.Common.Api.PreviewRoute
-{
-	[AttributeUsage(AttributeTargets.Class, AllowMultiple = false, Inherited = true)]
-	public sealed class PreviewRouteAttribute : Attribute
-	{ }
-}
