+++ src/Atomic.OutputCache/Controllers/CachedRenderController.cs
@@ -4,12 +4,10 @@ using Microsoft.AspNetCore.OutputCaching;
 using Microsoft.Extensions.Logging;
 using Umbraco.Cms.Core.Web;
 using Umbraco.Cms.Web.Common.Controllers;
-using Atomic.Common.Api.PreviewRoute;
 
 namespace Atomic.OutputCache.Controllers;
 
 [OutputCache]
-[PreviewRoute]
 public class CachedRenderController : RenderController
 {
     public CachedRenderController(ILogger<RenderController> logger,
