+++ src/Atomic.OutputCache/Controllers/CachedUmbracoApiController.cs
@@ -1,10 +1,12 @@
-﻿using Atomic.Common.Api.PreviewRoute;
+﻿using Atomic.Common.Api.Multilanguage;
+using Atomic.Common.Api.Preview;
 using Microsoft.AspNetCore.OutputCaching;
 using Umbraco.Cms.Web.Common.Controllers;
 
 namespace Atomic.OutputCache.Controllers;
 
 [OutputCache]
-[PreviewRoute]
+[EnablePreviewRoute]
+[SetContextCulture]
 public class CachedUmbracoApiController : UmbracoApiController
 { }
